requirements.txt
# Core dependencies
eventpulse==0.1.0

# Data processing and analysis
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Machine Learning and NLP
torch>=1.11.0
transformers>=4.15.0
sentence-transformers>=2.2.0
lightgbm>=3.3.0
xgboost>=1.5.0

# Financial data
yfinance>=0.1.70
alpha-vantage>=2.3.1
polygon-api-client>=1.9.0
python-binance>=1.0.15

# News and social media
newsapi-python>=0.2.6
tweepy>=4.14.0
praw>=7.6.0
feedparser>=6.0.8
beautifulsoup4>=4.11.1

# Web scraping and HTTP
aiohttp>=3.8.0
requests>=2.27.0
selenium>=4.1.0

# Database and caching
redis>=4.1.0
aioredis>=2.0.1
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
pymongo>=4.0.0

# Message queues and streaming
kafka-python>=2.0.2
pika>=1.2.0
celery>=5.2.0

# API and web frameworks
fastapi>=0.75.0
streamlit>=1.8.0
uvicorn>=0.17.0
gunicorn>=20.1.0

# Configuration and environment
pydantic>=1.9.0
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
dynaconf>=3.1.7

# Monitoring and logging
prometheus-client>=0.14.1
sentry-sdk>=1.5.8
structlog>=22.1.0
loguru>=0.6.0

# Testing and development
pytest>=7.1.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
black>=22.3.0
isort>=5.10.0
flake8>=4.0.0

# Security and authentication
cryptography>=36.0.0
PyJWT>=2.3.0
passlib>=1.7.4

# Deployment and containerization
docker>=5.0.0
kubernetes>=23.3.0

# Cloud services
boto3>=1.21.0
google-cloud-storage>=2.1.0
azure-storage-blob>=12.10.0

# Quantitative finance
quantlib>=1.26
zipline-reloaded>=2.2.0
pyfolio>=0.9.2
empyrical>=0.5.5

# Technical analysis
TA-Lib>=0.4.24
ta>=0.10.2

# Alternative data sources
spacy>=3.4.0
nltk>=3.7

# Production utilities
schedule>=1.1.0
click>=8.0.0
tqdm>=4.64.0
python-dateutil>=2.8.2 